backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.1
langchain-community==0.3.1
langchain-google-genai>=1.0.8
pymysql==1.1.1
chromadb==0.5.11
python-dotenv==1.0.1
pandas==2.2.3
sqlalchemy==2.0.35
